Changed FluidMenuBarExtra to be a struct that conforms to the Scene protocol, so it can be used just like the Apple MenuBarExtra SwiftUI type.
The state is stored via @StateObject using a glue struct (FluidMenuBarExtraStatusItemWrapper) to the FluidMenuBarExtraStatusItem.  This ensures the FluidMenuBarExtraStatusItem instance isn't thrown away and reinstantiated needlessly (though as per usual in SwiftUI, users of FluidMenuBarExtra will need to be careful not to change things between view structure refreshes in a way that makes SwiftUI throw away & recreate the underlying state).

Part of this - technically orthogonal but embroiled in my changes via my experimentation - is a refactor of how the image is handled, using a new Image enum in FluidMenuBarExtraStatusItem that eliminates the need for multiple init methods (in FluidMenuBarExtraStatusItem; FluidMenuBarExtra still has multiple because it has to mimic the MenuBarExtra API, warts and all).