Added the ability to control the visibility of the menubar item exactly like MenuBarExtra does - through an `isInserted` binding that can be passed to FluidMenuBarExtra.
This works both ways, too, through the binding - if the user removes the menubar item by command-dragging it out of the menubar, the bind value will be suitably updated to 'false'.

This was annoyingly difficult to get working.  SwiftUI's bindings are a real pain in the arse.  A key discovery was that they are partially transient - you cannot persist them beyond the lifetime of the View structures and expect them to work properly - they read as stuck on whatever value they had when the View structures were tossed out.  Curiously (and thankfully) they do still work to _update_ the value, even though you can't read the [correct] current value.  Bizarre.