Upgraded the demo app to a real GUI application, since Swift packages don't really support SwiftUI apps.
Previously the app had no menubar and couldn't become main properly, so the fudged 'settings' window would by default appear _behind_ Xcode's frontmost window.  There were some other glitches regarding the menubar item itself, too.

As part of this I also added some more content to the demonstration pop-up, including a dynamic element of adding & removing buttons to demonstrate window resizing.  This revealed a major issue with dynamic content, which - after a stupidly large number of hours of debugging, because SwiftUI is so inscrutable and poorly documented - I finally concluded that there's no way to directly fix the issue, or even figure out precisely why it occurs, but there is a workaround which is viable and realistic for most real-world apps (putting the meat of the SwiftUI code in a separate View struct, from wherever you declare the FluidMenuBarExtra).  Sigh.