The "alignRight" attribute now updates, including live - the window will slide smoothly to its new position as appropriate.
Updated the demo app to include a checkbox for this attribute, in order to play with it (and test it).

As a side-effect I simplified the logic in setWindowPosition(…).  I also added a guard that the new frame actually differs from the old frame before calling NSWindow methods to change it - possibly they ignore pointless resetting of the frame to its current value, but I don't want to rely on that (and them never having subtle side-effects).