The animation attribute now actually updates when changed after first initialisation.
Added controls for it to the demo app.

It doesn't update "live" as such because it has no visible effect on the pop-up window when it's already displayed.  You have to close the pop-up and re-open it to see the change in animation behaviour.  This is entirely intentional - re-showing the already visible window, just to do the opening animation, would be super annoying to an end-user.