Changes to the menu overall are now reflected (meaning going from no menu to a menu or vice versa, or between different NSMenu instances).
This is not fully "live" - if the menu is changed while it's being displayed, nothing changes visually.  This is somewhat intentional - changing the _contents_ of the menu is one thing, and left to NSMenu to handle in its normal way, but replacing an entire NSMenu instance with another (or switching to/from nil) has less clear expectations.  And in any case NSMenu should probably be replaced with Menu soon…