Added more control over the pop-up window's alignment relative to the menubar item - whether to [aim for] left, centred, or right alignment.
This introduces the PopUpAlignment enum.

As part of this I renamed the `alignRight` attribute to `screenClippingBehaviour`, and it now uses a new enum (ScreenClippingBehaviour) rather than being just a boolean.  Possibly in future there'll be other options than just the two current ones (e.g. I can vaguely see an argument for a "nextBestAlignment" option).