Clipping against the left edge of the screen is now handled correctly too (in addition to the right edge).
Extended the demo app with a box that can be incrementally enlarged or shrunk in the horizontal dimension, in order to demonstrate & test this functionality.

Incidentally, removed the overarching Form wrapper from the Settings view, as it was causing weird layout.  Now only a subset of the controls are embedded in a Form - and it now applies to the menubar item's popup too, not just the Settings window.